Fix mypy issues (#12)
* For now use mypy 1.13.0

* Adjustment for mypy, no change in code logic

* Add back type ignore

* Replace np.float_ by np.float32, as the former has been removed since NumPy 2.0

* Fix union type

* Type ignore

* Revert "Type ignore"

This reverts commit 52d5152a220ca1ab548194e05b3a02c671d5c048.

* Try float type

* Try np.generic

* Type ignore

* Type ignore

* Type ignore

* Apply pylint differently

* Skip type ignore

* Use floatarray instead of anyarray

* Use mypy 1.12.1

* No type ignore

* NDFloatArray = NDAnyArray

+++ examples/kaggle/run_tf_dataset_kaggle_example.py
@@ -19,7 +19,7 @@ import pandas as pd
 import targetran.tf as tt
 import tensorflow as tf
 
-NDAnyArray = np.typing.NDArray[np.float_]
+NDAnyArray = np.typing.NDArray[np.float32]
 
 # This will be the data path when you use "Add Data" on the right panel
 # of a Kaggle Notebook.
